requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
torch>=2.0.0
diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=7.0.0

# SaaS Features
reportlab>=4.0.0
requests>=2.31.0
python-multipart>=0.0.6
